+++ app/src/main/java/com/modern/btourist/Database/User.kt
@@ -3,7 +3,7 @@ package com.modern.btourist.Database
 data class User(var firstName: String,
                 var lastName: String,
                 var email: String,
-                var phone: Int,
+                var phone: Long,
                 var interest1: String,
                 var interest2: String,
                 var interest3: String,
@@ -15,7 +15,7 @@ data class User(var firstName: String,
                 var latitude: Double,
                 var longitude: Double){
 
-    constructor():this("","","",0,"","","",0,"","","",null,
+    constructor():this("","","",0L,"","","",0,"","","",null,
         0.0,0.0)
 
 }
